refactor shared fixtures to conftest and mark doc with test type
